Extract Find-related actions from CourseChoicesController
These actions relate specifically to guiding the user back to Find if
they haven't yet decided on a course. Move these into their own
controller, namespaced under the CourseChoices module.
