+++ app/views/candidate_interface/course_choices/have_you_chosen/ask.html.erb
