+++ app/views/candidate_interface/course_choices/have_you_chosen/go_to_find.html.erb
